[Desktop] Add the menu (always open) when the screen is very large.
PATH=/etapes
For #329
